requirements.txt

aiohttp>=3.11.13
beautifulsoup4>=4.13.3
google-api-python-client>=2.164.0
nest-asyncio>=1.6.0
pillow>=11.1.0
python-dotenv>=1.0.1
python-telegram-bot>=21.11.1
requests>=2.32.3
yt-dlp>=2025.2.19
binpan
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
telegram
trafilatura
aiohttp
beautifulsoup4
google-api-python-client
pillow
python-dotenv
python-telegram-bot
yt-dlp
aiohttp
beautifulsoup4
google-api-python-client
nest-asyncio
pillow
python-dotenv
python-telegram-bot
requests
trafilatura
yt-dlp
aiohttp
nest-asyncio
pillow
python-dotenv
python-telegram-bot
requests
yt-dlp
beautifulsoup4
google-api-python-client
trafilatura
yt-dlp
pillow
google-api-python-client
openai
